Added doxygen configuration ad output files
doxygen is a very good tool to document the code. I have added the configuration file to generate doxygen output in html as well as Latex. We can improve the documentation by adding structured comments in the code. For examples see https://www.stack.nl/~dimitri/doxygen/manual/docblocks.html

It is very easy to add comments and document the code nicely.